Rewrite comper decompression routine in sh2 asm

+++ src/include/mars/types.h
@@ -7,15 +7,15 @@
 #define __TYPES_H__
 
 
-typedef unsigned char   u8;
-typedef unsigned short  u16;
-typedef unsigned int    u32;
-typedef unsigned long   u64;
+typedef unsigned char       u8;
+typedef unsigned short      u16;
+typedef unsigned int        u32;
+typedef unsigned long long  u64;
 
-typedef signed char     s8;
-typedef signed short    s16;
-typedef signed int      s32;
-typedef signed long     s64;
+typedef signed char         s8;
+typedef signed short        s16;
+typedef signed int          s32;
+typedef signed long long    s64;
 
 
 #endif
